# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/lessons": {
      "get": {
        "description": "Returns every non-deleted lesson with startsAtUtc in [from, to).",
        "operationId": "LessonsController_list",
        "parameters": [
          {
            "name": "from",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": true,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "enrollmentId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SCHEDULED",
                "COMPLETED",
                "CANCELLED_CHARGED",
                "CANCELLED_UNCHARGED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List lessons in a time window (calendar feed)",
        "tags": [
          "scheduling"
        ]
      },
      "post": {
        "description": "startsAt accepts several dates for bulk creation. Overlapping the teacher returns 409 SCHEDULE_CONFLICT unless force=true.",
        "operationId": "LessonsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonListDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create one or many one-off lessons",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/lessons/{lessonId}/reschedule": {
      "patch": {
        "description": "scope=this detaches and moves a single lesson; scope=this_and_following shifts the series time onward. Conflicts return 409 unless force=true.",
        "operationId": "LessonsController_reschedule",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleLessonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reschedule a lesson",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/lessons/{lessonId}/status": {
      "patch": {
        "description": "Enforces the lesson state machine. Cancelling requires cancelledBy. No ledger effect in Stage 3.",
        "operationId": "LessonsController_transition",
        "parameters": [
          {
            "name": "lessonId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionLessonDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change a lesson status",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/lesson-series": {
      "get": {
        "description": "Paginated; state=deleted|all is owner-only.",
        "operationId": "SeriesController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "exclusiveMinimum": false,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "enrollmentId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "default": "active",
              "type": "string",
              "enum": [
                "active",
                "deleted",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonSeriesListDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List recurring lesson patterns",
        "tags": [
          "scheduling"
        ]
      },
      "post": {
        "operationId": "SeriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLessonSeriesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonSeriesDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a recurring pattern and materialize its lessons",
        "tags": [
          "scheduling"
        ]
      }
    },
    "/api/lesson-series/{seriesId}": {
      "get": {
        "operationId": "SeriesController_getDetail",
        "parameters": [
          {
            "name": "seriesId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonSeriesDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a recurring pattern",
        "tags": [
          "scheduling"
        ]
      },
      "patch": {
        "description": "Changing a schedule field regenerates future SCHEDULED lessons; a price-only change leaves existing lessons untouched.",
        "operationId": "SeriesController_update",
        "parameters": [
          {
            "name": "seriesId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLessonSeriesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LessonSeriesDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a recurring pattern",
        "tags": [
          "scheduling"
        ]
      },
      "delete": {
        "description": "Soft-deletes the pattern and its future scheduled lessons.",
        "operationId": "SeriesController_remove",
        "parameters": [
          {
            "name": "seriesId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a recurring pattern",
        "tags": [
          "scheduling"
        ]
      }
    }
  },
  "info": {
    "title": "Tutorio API",
    "description": "Financial calendar for private tutors",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
...
      "LessonListDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "workspaceId": {
                  "type": "string",
                  "format": "uuid"
                },
                "enrollmentId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "groupId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "seriesId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "packageId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "teacherId": {
                  "type": "string",
                  "format": "uuid"
                },
                "startsAtUtc": {
                  "type": "string",
                  "format": "date-time"
                },
                "durationMin": {
                  "type": "integer",
                  "minimum": 5,
                  "exclusiveMinimum": false,
                  "maximum": 720,
                  "exclusiveMaximum": false
                },
                "priceMinor": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": false,
                  "maximum": 2147483647,
                  "exclusiveMaximum": false
                },
                "currency": {
                  "type": "string",
                  "enum": [
                    "EUR",
                    "UAH",
                    "PLN",
                    "USD",
                    "GBP"
                  ]
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "SCHEDULED",
                    "COMPLETED",
                    "CANCELLED_CHARGED",
                    "CANCELLED_UNCHARGED"
                  ]
                },
                "isDetached": {
                  "type": "boolean"
                },
                "cancelledBy": {
                  "type": "string",
                  "enum": [
                    "TEACHER",
                    "STUDENT",
                    "GROUP"
                  ],
                  "nullable": true
                },
                "cancelledReason": {
                  "type": "string",
                  "nullable": true
                },
                "cancelledAt": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                },
                "completedAt": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                },
                "notes": {
                  "type": "string",
                  "nullable": true
                },
                "student": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "fullName": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "fullName"
                  ],
                  "nullable": true
                },
                "group": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "name"
                  ],
                  "nullable": true
                },
                "teacher": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "color": {
                      "type": "string",
                      "nullable": true
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "color"
                  ]
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "updatedAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "deletedAt": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "workspaceId",
                "enrollmentId",
                "groupId",
                "seriesId",
                "packageId",
                "teacherId",
                "startsAtUtc",
                "durationMin",
                "priceMinor",
                "currency",
                "status",
                "isDetached",
                "cancelledBy",
                "cancelledReason",
                "cancelledAt",
                "completedAt",
                "notes",
                "student",
                "group",
                "teacher",
                "createdAt",
                "updatedAt",
                "deletedAt"
              ]
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CreateLessonDto": {
        "type": "object",
        "properties": {
          "enrollmentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "groupId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "teacherId": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "type": "array",
            "minItems": 1,
            "maxItems": 50,
            "items": {
              "type": "string",
              "format": "date-time"
            }
          },
          "durationMin": {
            "type": "integer",
            "minimum": 5,
            "exclusiveMinimum": false,
            "maximum": 720,
            "exclusiveMaximum": false
          },
          "priceMinor": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 2147483647,
            "exclusiveMaximum": false
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "UAH",
              "PLN",
              "USD",
              "GBP"
            ]
          },
          "notes": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          }
        },
        "required": [
          "teacherId",
          "startsAt",
          "durationMin",
          "priceMinor",
          "currency"
        ]
      },
      "RescheduleLessonDto": {
        "type": "object",
        "properties": {
          "startsAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "durationMin": {
            "type": "integer",
            "minimum": 5,
            "exclusiveMinimum": false,
            "maximum": 720,
            "exclusiveMaximum": false
          },
          "scope": {
            "type": "string",
            "enum": [
              "this",
              "this_and_following"
            ],
            "default": "this"
          }
        },
        "required": [
          "startsAtUtc"
        ]
      },
      "LessonDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "enrollmentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "groupId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "seriesId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "packageId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "teacherId": {
            "type": "string",
            "format": "uuid"
          },
          "startsAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "durationMin": {
            "type": "integer",
            "minimum": 5,
            "exclusiveMinimum": false,
            "maximum": 720,
            "exclusiveMaximum": false
          },
          "priceMinor": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "maximum": 2147483647,
            "exclusiveMaximum": false
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "UAH",
              "PLN",
              "USD",
              "GBP"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED_CHARGED",
              "CANCELLED_UNCHARGED"
            ]
          },
          "isDetached": {
            "type": "boolean"
          },
          "cancelledBy": {
            "type": "string",
            "enum": [
              "TEACHER",
              "STUDENT",
              "GROUP"
            ],
            "nullable": true
          },
          "cancelledReason": {
            "type": "string",
            "nullable": true
          },
          "cancelledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "student": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "fullName": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "fullName"
            ],
            "nullable": true
          },
          "group": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ],
            "nullable": true
          },
          "teacher": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              },
              "color": {
                "type": "string",
                "nullable": true
              }
            },
            "required": [
              "id",
              "name",
              "color"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "workspaceId",
          "enrollmentId",
          "groupId",
          "seriesId",
          "packageId",
          "teacherId",
          "startsAtUtc",
          "durationMin",
          "priceMinor",
          "currency",
          "status",
          "isDetached",
          "cancelledBy",
          "cancelledReason",
          "cancelledAt",
          "completedAt",
          "notes",
          "student",
          "group",
          "teacher",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "TransitionLessonDto": {
        "type": "object",
        "properties": {
          "targetStatus": {
            "type": "string",
            "enum": [
              "SCHEDULED",
              "COMPLETED",
              "CANCELLED_CHARGED",
              "CANCELLED_UNCHARGED"
            ]
          },
          "cancelledBy": {
            "type": "string",
            "enum": [
              "TEACHER",
              "STUDENT",
              "GROUP"
            ]
          },
          "cancelledReason": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          }
        },
        "required": [
          "targetStatus"
        ]
      },
...
    }
  }
}